Move FairfaxFilesProcessor to processor package.
- Move FairfaxFilesProcessor.groovy to processor package.
- Move ParentGroupingProcessor.groovy and SipForFolderProcessor.groovy to
  processor.type package.
