+++ core/src/test/groovy/nz/govt/natlib/tools/sip/generation/fairfax/scenarios/NoMatchingSipDefinitionTest.groovy
@@ -1,15 +1,12 @@
 package nz.govt.natlib.tools.sip.generation.fairfax.scenarios
 
 import groovy.util.logging.Slf4j
-import nz.govt.natlib.tools.sip.IEEntityType
-import nz.govt.natlib.tools.sip.extraction.SipXmlExtractor
 import nz.govt.natlib.tools.sip.generation.fairfax.FairfaxFile
-import nz.govt.natlib.tools.sip.generation.fairfax.FairfaxFilesProcessor
+import nz.govt.natlib.tools.sip.generation.fairfax.processor.FairfaxFilesProcessor
 import nz.govt.natlib.tools.sip.generation.fairfax.FairfaxProcessingParameters
 import nz.govt.natlib.tools.sip.generation.fairfax.TestHelper
 import nz.govt.natlib.tools.sip.generation.fairfax.TestHelper.TestMethodState
 import nz.govt.natlib.tools.sip.generation.fairfax.parameters.ProcessingType
-import nz.govt.natlib.tools.sip.state.SipProcessingExceptionReasonType
 import org.junit.Before
 import org.junit.Ignore
 import org.junit.Test
@@ -20,7 +17,6 @@ import java.nio.file.Path
 import java.time.LocalDate
 
 import static org.hamcrest.core.Is.is
-import static org.junit.Assert.assertFalse
 import static org.junit.Assert.assertThat
 import static org.junit.Assert.assertTrue
